将返回的信息，统一成Response类之后的程序，尚未postman测试接口方法

+++ mpmarket/src/main/java/com/mpprojects/mpmarket/model/shop/CouponRule.java
@@ -23,6 +23,6 @@ public class CouponRule {
     private Long id;
 
     private String ruleName;        //规则的名称
-    private Long ruleNumber;        //规则类型的编号，见上面。
+    private Integer ruleNumber;        //规则类型的编号，见上面。
     private String description;
 }
